Move model to project models file and add migration
